Restructure `git` crate
